backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-generativeai==0.3.2
transformers==4.35.2
torch==2.7.1
googlemaps==4.10.0
requests==2.31.0
langdetect==1.0.9
SpeechRecognition==3.10.0
pydub==0.25.1
gtts==2.4.0
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.0
regex==2023.10.3
gunicorn==21.2.0